[v0.87.1][maintainability] Split pr_cmd.rs by responsibility
