+++ adl/src/cli/pr_cmd/doctor.rs
@@ -1,0 +1,384 @@
+use super::*;
+use serde::Serialize;
+use std::fs;
+use std::path::{Path, PathBuf};
+
+#[derive(Debug, Clone, PartialEq, Eq, Serialize)]
+struct DoctorPreflightJsonPullRequest {
+    number: u32,
+    head_ref_name: String,
+    state: &'static str,
+    url: String,
+}
+
+#[derive(Debug, Clone, PartialEq, Eq)]
+struct DoctorPreflightResult {
+    open_pr_count: usize,
+    open_prs: Vec<DoctorPreflightJsonPullRequest>,
+    status: &'static str,
+}
+
+#[derive(Debug, Clone, PartialEq, Eq)]
+struct DoctorReadyResult {
+    lifecycle_state: &'static str,
+    worktree: Option<String>,
+    source: String,
+    root_stp: String,
+    root_input: String,
+    root_output: String,
+    wt_stp: Option<String>,
+    wt_input: Option<String>,
+    wt_output: Option<String>,
+    status: &'static str,
+}
+
+#[derive(Debug, Clone, PartialEq, Eq, Serialize)]
+struct DoctorJsonOutput {
+    schema: &'static str,
+    issue: u32,
+    version: String,
+    slug: String,
+    branch: String,
+    mode: &'static str,
+    preflight_status: &'static str,
+    open_pr_count: usize,
+    open_prs: Vec<DoctorPreflightJsonPullRequest>,
+    lifecycle_state: Option<&'static str>,
+    ready_status: Option<&'static str>,
+    worktree: Option<String>,
+    source: Option<String>,
+    root_stp: Option<String>,
+    root_input: Option<String>,
+    root_output: Option<String>,
+    wt_stp: Option<String>,
+    wt_input: Option<String>,
+    wt_output: Option<String>,
+    doctor_status: &'static str,
+}
+
+pub(super) fn run_doctor(parsed: DoctorArgs, label: &str) -> Result<()> {
+    let repo_root = primary_checkout_root()?;
+    let repo = default_repo(&repo_root)?;
+    let (version, slug) = resolve_doctor_scope_and_slug(&repo_root, &parsed, label)?;
+    let issue_ref = IssueRef::new(parsed.issue, version.clone(), slug.clone())?;
+    let branch = issue_ref.branch_name("codex");
+
+    let preflight = run_doctor_preflight(&repo, &version, &branch)?;
+    let ready = match parsed.mode {
+        DoctorMode::Preflight => None,
+        DoctorMode::Ready | DoctorMode::Full => {
+            Some(run_doctor_ready(&repo_root, &repo, &issue_ref, &branch)?)
+        }
+    };
+    let mode = doctor_mode_name(&parsed.mode);
+    let doctor_status = match parsed.mode {
+        DoctorMode::Preflight => preflight.status,
+        DoctorMode::Ready => ready.as_ref().map(|x| x.status).unwrap_or("BLOCK"),
+        DoctorMode::Full => {
+            if preflight.status == "PASS" && ready.as_ref().map(|x| x.status) == Some("PASS") {
+                "PASS"
+            } else {
+                "BLOCK"
+            }
+        }
+    };
+
+    if parsed.json {
+        print_json(&DoctorJsonOutput {
+            schema: "adl.pr.doctor.v1",
+            issue: parsed.issue,
+            version,
+            slug,
+            branch,
+            mode,
+            preflight_status: preflight.status,
+            open_pr_count: preflight.open_pr_count,
+            open_prs: preflight.open_prs,
+            lifecycle_state: ready.as_ref().map(|x| x.lifecycle_state),
+            ready_status: ready.as_ref().map(|x| x.status),
+            worktree: ready.as_ref().and_then(|x| x.worktree.clone()),
+            source: ready.as_ref().map(|x| x.source.clone()),
+            root_stp: ready.as_ref().map(|x| x.root_stp.clone()),
+            root_input: ready.as_ref().map(|x| x.root_input.clone()),
+            root_output: ready.as_ref().map(|x| x.root_output.clone()),
+            wt_stp: ready.as_ref().and_then(|x| x.wt_stp.clone()),
+            wt_input: ready.as_ref().and_then(|x| x.wt_input.clone()),
+            wt_output: ready.as_ref().and_then(|x| x.wt_output.clone()),
+            doctor_status,
+        })?;
+    } else {
+        println!("ISSUE={}", parsed.issue);
+        println!("VERSION={version}");
+        println!("SLUG={slug}");
+        println!("BRANCH={branch}");
+        print_doctor_preflight_text(&preflight);
+        if let Some(ready) = &ready {
+            print_doctor_ready_text(ready);
+        }
+        println!("DOCTOR_MODE={mode}");
+        println!("DOCTOR_STATUS={doctor_status}");
+    }
+    Ok(())
+}
+
+fn resolve_doctor_scope_and_slug(
+    repo_root: &Path,
+    parsed: &DoctorArgs,
+    label: &str,
+) -> Result<(String, String)> {
+    if let (Some(version), Some(slug)) = (parsed.version.clone(), parsed.slug.clone()) {
+        return Ok((version, slug));
+    }
+    let inferred = resolve_issue_scope_and_slug_from_local_state(repo_root, parsed.issue)?;
+    let version = parsed
+        .version
+        .clone()
+        .or(inferred.as_ref().map(|x| x.0.clone()))
+        .unwrap_or_else(|| DEFAULT_VERSION.to_string());
+    let slug = match parsed.mode {
+        DoctorMode::Preflight => parsed
+            .slug
+            .clone()
+            .or(inferred.map(|x| x.1))
+            .unwrap_or_else(|| format!("issue-{}", parsed.issue)),
+        DoctorMode::Ready | DoctorMode::Full => {
+            parsed.slug.clone().or(inferred.map(|x| x.1)).ok_or_else(|| {
+                if label == "ready" {
+                    anyhow!("ready: could not infer slug; pass --slug or run start first")
+                } else {
+                    anyhow!(
+                        "doctor: could not infer slug for readiness check; pass --slug or create the execution context first"
+                    )
+                }
+            })?
+        }
+    };
+    Ok((version, slug))
+}
+
+fn run_doctor_preflight(repo: &str, version: &str, branch: &str) -> Result<DoctorPreflightResult> {
+    let unresolved = unresolved_milestone_pr_wave(repo, version, Some(branch))?;
+    let open_prs = unresolved
+        .iter()
+        .map(|pr| DoctorPreflightJsonPullRequest {
+            number: pr.number,
+            head_ref_name: pr.head_ref_name.clone(),
+            state: if pr.is_draft { "draft" } else { "ready" },
+            url: pr.url.clone(),
+        })
+        .collect::<Vec<_>>();
+    if open_prs.is_empty() {
+        Ok(DoctorPreflightResult {
+            open_pr_count: 0,
+            open_prs,
+            status: "PASS",
+        })
+    } else {
+        Ok(DoctorPreflightResult {
+            open_pr_count: open_prs.len(),
+            open_prs,
+            status: "BLOCK",
+        })
+    }
+}
+
+fn run_doctor_ready(
+    repo_root: &Path,
+    repo: &str,
+    issue_ref: &IssueRef,
+    branch: &str,
+) -> Result<DoctorReadyResult> {
+    let worktree_path = issue_ref.default_worktree_path(
+        repo_root,
+        std::env::var_os("ADL_WORKTREE_ROOT")
+            .map(PathBuf::from)
+            .as_deref(),
+    );
+    let source_path = resolve_issue_prompt_path(repo_root, issue_ref)?;
+    let root_stp = issue_ref.task_bundle_stp_path(repo_root);
+    let wt_stp = issue_ref.task_bundle_stp_path(&worktree_path);
+    let root_bundle_input = issue_ref.task_bundle_input_path(repo_root);
+    let root_bundle_output = issue_ref.task_bundle_output_path(repo_root);
+    let wt_bundle_input = issue_ref.task_bundle_input_path(&worktree_path);
+    let wt_bundle_output = issue_ref.task_bundle_output_path(&worktree_path);
+    let closed_completed =
+        lifecycle::issue_is_closed_and_completed(issue_ref.issue_number(), repo)?;
+
+    validate_issue_prompt_exists(&source_path)?;
+    validate_bootstrap_stp(repo_root, &source_path)?;
+    validate_authored_prompt_surface("doctor", &source_path, PromptSurfaceKind::IssuePrompt)?;
+    if closed_completed {
+        lifecycle::reconcile_closed_completed_issue_bundle(
+            repo_root,
+            issue_ref,
+            &root_bundle_output,
+        )?;
+    }
+    if !root_stp.is_file() {
+        bail!("doctor: missing root stp: {}", root_stp.display());
+    }
+    validate_bootstrap_stp(repo_root, &root_stp)?;
+    validate_authored_prompt_surface("doctor", &root_stp, PromptSurfaceKind::Stp)?;
+    if closed_completed {
+        validate_initialized_cards(
+            issue_ref.issue_number(),
+            issue_ref.slug(),
+            &root_bundle_input,
+            &root_bundle_output,
+        )?;
+        return Ok(DoctorReadyResult {
+            lifecycle_state: "closed",
+            worktree: None,
+            source: path_relative_to_repo(repo_root, &source_path),
+            root_stp: path_relative_to_repo(repo_root, &root_stp),
+            root_input: path_relative_to_repo(repo_root, &root_bundle_input),
+            root_output: path_relative_to_repo(repo_root, &root_bundle_output),
+            wt_stp: None,
+            wt_input: None,
+            wt_output: None,
+            status: "PASS",
+        });
+    }
+    validate_initialized_cards(
+        issue_ref.issue_number(),
+        issue_ref.slug(),
+        &root_bundle_input,
+        &root_bundle_output,
+    )?;
+    let root_input_body = fs::read_to_string(&root_bundle_input).with_context(|| {
+        format!(
+            "doctor: read root input card: {}",
+            root_bundle_input.display()
+        )
+    })?;
+    let root_branch = field_line_value(&root_bundle_input, "Branch")?;
+    let root_indicates_pre_run = branch_indicates_unbound_state(&root_branch)
+        || root_input_body.contains(
+            "This issue is not started yet; do not assume a branch or worktree already exists.",
+        )
+        || root_input_body
+            .contains("Do not assume a branch or worktree already exists before `pr run`.");
+    if !worktree_path.is_dir() {
+        if root_indicates_pre_run {
+            return Ok(DoctorReadyResult {
+                lifecycle_state: "pre_run",
+                worktree: None,
+                source: path_relative_to_repo(repo_root, &source_path),
+                root_stp: path_relative_to_repo(repo_root, &root_stp),
+                root_input: path_relative_to_repo(repo_root, &root_bundle_input),
+                root_output: path_relative_to_repo(repo_root, &root_bundle_output),
+                wt_stp: None,
+                wt_input: None,
+                wt_output: None,
+                status: "PASS",
+            });
+        }
+        bail!("doctor: missing worktree: {}", worktree_path.display());
+    }
+    if root_indicates_pre_run
+        && (!wt_stp.is_file() || !wt_bundle_input.is_file() || !wt_bundle_output.is_file())
+    {
+        return Ok(DoctorReadyResult {
+            lifecycle_state: "pre_run",
+            worktree: None,
+            source: path_relative_to_repo(repo_root, &source_path),
+            root_stp: path_relative_to_repo(repo_root, &root_stp),
+            root_input: path_relative_to_repo(repo_root, &root_bundle_input),
+            root_output: path_relative_to_repo(repo_root, &root_bundle_output),
+            wt_stp: None,
+            wt_input: None,
+            wt_output: None,
+            status: "PASS",
+        });
+    }
+    let wt_branch = run_capture(
+        "git",
+        &[
+            "-C",
+            path_str(&worktree_path)?,
+            "rev-parse",
+            "--abbrev-ref",
+            "HEAD",
+        ],
+    )?;
+    if wt_branch.trim() != branch {
+        bail!(
+            "doctor: worktree branch mismatch for {}",
+            worktree_path.display()
+        );
+    }
+    if !wt_stp.is_file() {
+        bail!("doctor: missing worktree stp: {}", wt_stp.display());
+    }
+    validate_bootstrap_stp(&worktree_path, &wt_stp)?;
+    validate_authored_prompt_surface("doctor", &wt_stp, PromptSurfaceKind::Stp)?;
+    validate_ready_cards(
+        repo_root,
+        issue_ref.issue_number(),
+        issue_ref.slug(),
+        wt_branch.trim(),
+        &root_bundle_input,
+        &root_bundle_output,
+    )?;
+    validate_ready_cards(
+        &worktree_path,
+        issue_ref.issue_number(),
+        issue_ref.slug(),
+        wt_branch.trim(),
+        &wt_bundle_input,
+        &wt_bundle_output,
+    )?;
+
+    Ok(DoctorReadyResult {
+        lifecycle_state: "run_bound",
+        worktree: Some(path_relative_to_repo(repo_root, &worktree_path)),
+        source: path_relative_to_repo(repo_root, &source_path),
+        root_stp: path_relative_to_repo(repo_root, &root_stp),
+        root_input: path_relative_to_repo(repo_root, &root_bundle_input),
+        root_output: path_relative_to_repo(repo_root, &root_bundle_output),
+        wt_stp: Some(path_relative_to_repo(repo_root, &wt_stp)),
+        wt_input: Some(path_relative_to_repo(repo_root, &wt_bundle_input)),
+        wt_output: Some(path_relative_to_repo(repo_root, &wt_bundle_output)),
+        status: "PASS",
+    })
+}
+
+fn doctor_mode_name(mode: &DoctorMode) -> &'static str {
+    match mode {
+        DoctorMode::Full => "full",
+        DoctorMode::Ready => "ready",
+        DoctorMode::Preflight => "preflight",
+    }
+}
+
+fn print_doctor_preflight_text(preflight: &DoctorPreflightResult) {
+    println!("OPEN_PR_COUNT={}", preflight.open_pr_count);
+    for pr in &preflight.open_prs {
+        println!(
+            "OPEN_PR=#{}|{}|{}|{}",
+            pr.number, pr.head_ref_name, pr.state, pr.url
+        );
+    }
+    println!("PREFLIGHT={}", preflight.status);
+}
+
+fn print_doctor_ready_text(ready: &DoctorReadyResult) {
+    println!("LIFECYCLE_STATE={}", ready.lifecycle_state);
+    if let Some(worktree) = &ready.worktree {
+        println!("WORKTREE={worktree}");
+    }
+    println!("SOURCE={}", ready.source);
+    println!("ROOT_STP={}", ready.root_stp);
+    println!("ROOT_INPUT={}", ready.root_input);
+    println!("ROOT_OUTPUT={}", ready.root_output);
+    if let Some(wt_stp) = &ready.wt_stp {
+        println!("WT_STP={wt_stp}");
+    }
+    if let Some(wt_input) = &ready.wt_input {
+        println!("WT_INPUT={wt_input}");
+    }
+    if let Some(wt_output) = &ready.wt_output {
+        println!("WT_OUTPUT={wt_output}");
+    }
+    println!("READY={}", ready.status);
+}
